requirements.txt
behave
nose
requests
pillow